docs

+++ server/middleware/delete-queries/routine-delete-query.js
@@ -1,10 +1,15 @@
 const RoutineModel = require('../../models/Routine')
+const {BadRequestError} = require('../../errors')
 
 const routineDeleteQuery = (req, res, next) => {
 
     const {
         params: {routineId}
     } = req
+
+    if (!routineId) {
+        throw new BadRequestError('Missing routine id!')
+    }
 
     const query = RoutineModel.findByIdAndDelete(routineId)
